automate_upload.py
This will automatically upload videos to TikTok in chronological order that the video was created. Does not double upload videos already posted. It also lists the date the video was made in the title/caption for the TikTok post. 

To do, just paste sessionid and directory to the script and make sure the script is in the same directory as the ones as uploader.py